Initial Commit: Basic compile_commands generator
A basic working implementation of a compile commands DB generator in
Rust

This tool is designed to replace a workflow using `bear`[1] to generate a
compile_commands.json file and then `compdb`[2] to do insert discovered
headers into the compile commands (because bear does not do this).

This tool is heavily geared towards one specific use case right now,
which is the CRuby source code: a large, and old codebase, using a
complex autotools based workflow. The goal is to generate a sufficiently
complete compilation db, that I can use it in conjunction with clangd to
provide complete code navigation around the codebase. I'm testing this
in Emacs with lsp-mode.

rccgen runs `make` in dry-run mode, falling back to `make all` if the
first command errors. It then parses the output looking for compiler
commands, and processes them for inclusion into the json output.

It then tries to do header discovery by looking through all the
extracted compiler commands for -I flags, and includes those in the
generated output too.

Finally it writes the compile_commands.json to the working directory.

rccgen supports partial output, that is, when make dry-run returns a
non-zero exit code but did still output text that contained compiler
commands we'll capture those too.

We also try some basic validation on paths and shell commands that we
see, just checking they're not so long as to cause DOS attacks, and
stripping null bytes and command characters.

[1]: https://github.com/rizsotto/Bear
[2]: https://github.com/Sarcasm/compdb